accurate crossover points

+++ R_package/include/GenomicsCommon.h
@@ -3,8 +3,9 @@
 #include <vector>
 #include <string>
 #include <random>
+#include <cstdint>
 #include <Rcpp.h>
 
 namespace GenomicsCommon {
-	using Pos = long long;
+	using Pos = int64_t;
 }
